some various small bug fixes
git-svn-id: file:///home/scott/regretitron_publish/regretitron_filtered3_svn/trunk@164 56bb13ec-af9c-4a7e-874f-62681e7d7c12

+++ PokerPlayerMFC/strategy.cpp
@@ -143,7 +143,7 @@ Strategy::Strategy(string xmlfilename, bool preload, LoggerClass & strategylogge
 			if(thisoffset != nextoffset)
 				REPORT("redundant offset storage has revealed errors in strategy file");
 			dataoffset[gr][n-2] = thisoffset;
-			nextoffset = thisoffset + get_property(gettree(), maxorderindex_tag())[gr][n] * cardmach->getcardsimax(gr) * (n+1);
+			nextoffset = thisoffset + get_property(gettree(), maxorderindex_tag())[gr][n] * static_cast< int64 >( cardmach->getcardsimax(gr) ) * static_cast< int64 >(n+1);
 		}
 	}
 	if(dataoffset[0][7] != strategyfile.Tell() || strategyfilesize != nextoffset)
